Add the DATABASE_URL as an environment variable to deployment configs
The URL value will be stored as an entry in the database's secret
object and imported by the deployment configs that need it.

This removes the need to pass in the database user, name, and password
to each individual deployment. The host and port need to remain
so that these pods can wait for the database service to be available